refactored timespan to simplify future extensions
