+++ vis/js/reducers/index.js
@@ -2,54 +2,57 @@
  * The main reducer where all subreducers are combined into one.
  */
 import { combineReducers } from "redux";
-import zoom from "./zoom";
+
+import animation from "./animation";
+import areas from "./areas";
+import bubbleOrder from "./bubbleOrder";
+import chart from "./chart";
 import chartType from "./chartType";
-import localization from "./localization";
+import contextLine from "./contextLine";
+import data from "./data";
+import files from "./files";
 import heading from "./heading";
+import highlightedBubble from "./highlightedBubble";
+import hyphenationLang from "./hyphenationLang";
+import list from "./list";
+import localization from "./localization";
+import misc from "./misc";
+import modals from "./modals";
+import paperOrder from "./paperOrder";
+import previewModal from "./previewModal";
+import query from "./query";
 import selectedBubble from "./selectedBubble";
 import selectedPaper from "./selectedPaper";
-import query from "./query";
-import files from "./files";
-import contextLine from "./contextLine";
 import service from "./service";
-import list from "./list";
-import data from "./data";
-import previewModal from "./previewModal";
-import areas from "./areas";
-import chart from "./chart";
-import animation from "./animation";
-import highlightedBubble from "./highlightedBubble";
-import bubbleOrder from "./bubbleOrder";
-import paperOrder from "./paperOrder";
-import hyphenationLang from "./hyphenationLang";
-import modals from "./modals";
-import toolbar from "./toolbar";
-import misc from "./misc";
 import streamgraph from "./streamgraph";
+import timespan from "./timespan";
+import toolbar from "./toolbar";
+import zoom from "./zoom";
 
 export default combineReducers({
-  zoom,
-  query,
+  animation,
+  areas,
+  bubbleOrder,
+  chart,
   chartType,
+  contextLine,
+  data,
+  files,
+  heading,
+  highlightedBubble,
+  hyphenationLang,
+  list,
+  localization,
+  misc,
+  modals,
+  paperOrder,
+  previewModal,
+  query,
   selectedBubble,
   selectedPaper,
-  heading,
-  localization,
-  files,
-  contextLine,
   service,
-  list,
-  data,
-  previewModal,
-  areas,
-  chart,
-  animation,
-  highlightedBubble,
-  bubbleOrder,
-  paperOrder,
-  hyphenationLang,
-  modals,
-  toolbar,
-  misc,
   streamgraph,
+  timespan,
+  toolbar,
+  zoom,
 });
